+++ packages/orq-rc/RELEASES.md
@@ -9709,3 +9709,13 @@ Based on:
 - [typescript v4.10.0-rc.75] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.75] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.75 - packages/orq-rc
+
+## 2026-06-03 15:04:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.81] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.81] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.81 - packages/orq-rc